backup_python312_upgrade/requirement.txt
# Discord ADR Bot v1.5 - 依賴套件清單

# 核心 Discord 套件
discord.py>=2.5.2
discord.py[voice]>=2.5.2

# 資料庫相關
aiosqlite>=0.21.0

# 圖片處理
Pillow>=11.2.1

# 環境變數管理
python-dotenv>=1.1.0

# 非同步 HTTP 客戶端
aiohttp>=3.11.18

# 效能優化（非 Windows）
uvloop>=0.19.0; sys_platform != "win32"

# Windows 彩色輸出支援
colorama>=0.4.6; sys_platform == "win32"

# 日誌輪轉
python-json-logger>=2.0.7

# 日期時間處理
python-dateutil>=2.8.2

# 正則表達式增強
regex>=2023.12.25

# 資料驗證
pydantic>=2.5.0

# 快取管理
cachetools>=5.3.2

# 檔案監控（可選）
watchdog>=3.0.0

# 壓縮支援
zstandard>=0.22.0

# 加密支援（可選）
cryptography>=41.0.0

# 網路工具
requests>=2.31.0
tldextract>=3.7.0

# 進度條（開發用）
tqdm>=4.66.0

# 測試框架（開發用）
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# 程式碼品質（開發用）
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0

# 文檔生成（開發用）
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0

# 部署工具（可選）
gunicorn>=21.2.0
supervisor>=4.2.5

# 監控工具（可選）
psutil>=5.9.0
prometheus-client>=0.19.0

# 開發工具（可選）
ipython>=8.18.0
jupyter>=1.0.0

# 安全性檢查（可選）
bandit>=1.7.5
safety>=2.3.0